combine reducers reformat
