+++ client/src/components/Cart/CartBody.js
@@ -4,7 +4,7 @@ import styled from "styled-components";
 import CartItemInput from "./CartItemInput";
 import CartItemDescription from "./CartItemDescription";
 
-import { getCartItemArray } from "../reducers/index";
+import { getCartItemArray } from "../reducers/cart-reducers";
 import { useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
