+++ client/src/components/Cart/CartFooter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { getCartItemArray } from "../reducers/index";
+import { getCartItemArray } from "../reducers/cart-reducers";
 import { useDispatch } from "react-redux";
 import { useSelector } from "react-redux";
 import { emptyCart } from "../action";
