+++ client/src/components/reducers/bodypart-reducers.js
@@ -1,0 +1,11 @@
+const initialState = {
+  sorting: 'All',
+};
+
+export default function bodyPartReducer(state = initialState, action) {
+  switch (action.type) {
+    default: {
+      return state;
+    }
+  }
+}
