+++ client/src/components/reducers/sort-reducers.js
@@ -1,0 +1,11 @@
+const initialState = {
+  sorting: 'Alphabetically (A to Z)',
+};
+
+export default function sortReducer(state = initialState, action) {
+  switch (action.type) {
+    default: {
+      return state;
+    }
+  }
+}
