+++ client/src/components/store.js
@@ -1,0 +1,13 @@
+import { createStore } from 'redux';
+
+import reducer from './reducers/index';
+
+export default function configureStore(initialState) {
+  const store = createStore(
+    reducer,
+    initialState,
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  );
+
+  return store;
+}
